pyproject.toml
[project]
name = "postgrid_integration"
authors = [
    { name = "Avunu LLC", email = "mail@avu.nu"}
]
description = "PostGrid Integration for Mailing Frappe Documents"
requires-python = ">=3.10"
readme = "README.md"
dynamic = ["version"]
dependencies = [
    "postgrid-python",
    "pyjwt",
    "pypdf"
]

[build-system]
requires = ["flit_core >=3.4,<4"]
build-backend = "flit_core.buildapi"

# These dependencies are only installed when developer mode is enabled
[tool.bench.dev-dependencies]
# package_name = "~=1.1.0"
